requirements.txt
openai
requests
python-dotenv
streamlit

pandas-profiling==3.2.0
ydata-profiling==4.16.1
PyYAML>=5.0.0
seaborn>=0.10.1
numba<=0.61,>=0.56.0
statsmodels>=0.13.2,<1
markupsafe==2.1.1
visions==0.7.4
matplotlib==3.10
numpy==1.21.5
